requirements.txt
fastapi
uvicorn
pytorch-lightning
opencv-python
albumentations
ultralytics
exifread
pillow